Moved EvalPrecedence to Entities folder
