+++ MathEvaluation/Entities/MathOperandOperator.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using MathEvaluation.Context;
 
 namespace MathEvaluation.Entities;
